Modified app.rb to check for exitstance of the name parameter.

+++ app.rb
@@ -1,4 +1,8 @@
 require_relative "dependencies"
+ActiveRecord::Base.establish_connection(
+  adapter: 'sqlite3',
+  database: 'test.sqlite3'
+)
 
 class App < Sinatra::Base
   get "/" do
@@ -11,16 +15,20 @@ class App < Sinatra::Base
   end
 
   get "/find/employee" do
-    name = params["name"]
-    employee = Employee.find_by(name: name)
-    employee.to_json
+    if params["name"] != nil
+      name = params["name"]
+      employee = Employee.find_by(name: name)
+      employee.to_json
+    end
   end
 
   delete "/delete/employee" do
-    name = params["name"]
-    employee = Employee.find_by(name: name)
-    employee.destroy
-    employee.to_json
+    if params["name"] != nil
+      name = params["name"]
+      employee = Employee.find_by(name: name)
+      employee.destroy
+      employee.to_json
+    end
   end
 
   post "/create/employee" do
@@ -30,11 +38,13 @@ class App < Sinatra::Base
   end
 
   patch "/update/employee" do
-    name = params["name"]
-    employee = Employee.find_by(name: name)
-    employee_data = JSON.parse(request.body.read)
-    employee.update(employee_data)
-    employee.to_json
+    if params["name"] != nil
+      name = params["name"]
+      employee = Employee.find_by(name: name)
+      employee_data = JSON.parse(request.body.read)
+      employee.update(employee_data)
+      employee.to_json
+    end
   end
 
   run! if app_file == $PROGRAM_NAME
